Add numpy and pandas, separate tess/tracking/service/ tests for skinny client initial tests. Override tracking uri fixture for service/ tests
