+++ mlflow/types/schema.py
@@ -1,19 +1,8 @@
 import json
 from enum import Enum
-import warnings
 
-try:
-    import numpy as np
-    has_np = True
-except ImportError as e:
-    warnings.warn("Schema options will not work as expected due to {}".format(e))
-    has_np = False
-
-try:
-    import pandas as pd
-except ImportError as e:
-    warnings.warn("Schema options will not work as expected due to {}".format(e))
-
+import numpy as np
+import pandas as pd
 from typing import Dict, Any, List, Union, Optional
 
 from mlflow.exceptions import MlflowException
@@ -43,21 +32,20 @@ class DataType(Enum):
     # integers and boolean values. We do not use them here for now as most downstream tools are
     # most likely to use / expect native numpy types and would not be compatible with the extension
     # types.
-    if has_np:
-        boolean = (1, np.dtype("bool"), "BooleanType")
-        """Logical data (True, False) ."""
-        integer = (2, np.dtype("int32"), "IntegerType")
-        """32b signed integer numbers."""
-        long = (3, np.dtype("int64"), "LongType")
-        """64b signed integer numbers. """
-        float = (4, np.dtype("float32"), "FloatType")
-        """32b floating point numbers. """
-        double = (5, np.dtype("float64"), "DoubleType")
-        """64b floating point numbers. """
-        string = (6, np.dtype("str"), "StringType", _pandas_string_type())
-        """Text data."""
-        binary = (7, np.dtype("bytes"), "BinaryType", np.object)
-        """Sequence of raw bytes."""
+    boolean = (1, np.dtype("bool"), "BooleanType")
+    """Logical data (True, False) ."""
+    integer = (2, np.dtype("int32"), "IntegerType")
+    """32b signed integer numbers."""
+    long = (3, np.dtype("int64"), "LongType")
+    """64b signed integer numbers. """
+    float = (4, np.dtype("float32"), "FloatType")
+    """32b floating point numbers. """
+    double = (5, np.dtype("float64"), "DoubleType")
+    """64b floating point numbers. """
+    string = (6, np.dtype("str"), "StringType", _pandas_string_type())
+    """Text data."""
+    binary = (7, np.dtype("bytes"), "BinaryType", np.object)
+    """Sequence of raw bytes."""
 
     def __repr__(self):
         return self.name
@@ -65,11 +53,11 @@ class DataType(Enum):
     def __str(self):
         return self.name
 
-    def to_numpy(self): # -> np.dtype, commented type info for lazy loading of numpy
+    def to_numpy(self) -> np.dtype:
         """Get equivalent numpy data type. """
         return self._numpy_type
 
-    def to_pandas(self): # -> np.dtype, commented type info for lazy loading of numpy
+    def to_pandas(self) -> np.dtype:
         """Get equivalent pandas data type. """
         return self._pandas_type
 
@@ -164,11 +152,11 @@ class Schema(object):
         """ Get column types of the columns in the dataset."""
         return [x.type for x in self._cols]
 
-    def numpy_types(self):  # -> List[np.dtype], commented type info for lazy loading of numpy
+    def numpy_types(self) -> List[np.dtype]:
         """ Convenience shortcut to get the datatypes as numpy types."""
         return [x.type.to_numpy() for x in self.columns]
 
-    def pandas_types(self):  # -> List[np.dtype], commented type info for lazy loading of numpy 
+    def pandas_types(self) -> List[np.dtype]:
         """ Convenience shortcut to get the datatypes as pandas types."""
         return [x.type.to_pandas() for x in self.columns]
